requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.27.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
neo4j==5.13.0
chromadb==0.4.17
numpy==1.26.0
pandas==2.1.1
pydantic==2.4.2
pydantic-settings==2.0.3
loguru==0.7.2
aiohttp==3.8.6
asyncio==3.4.3
httpx==0.25.0
python-dateutil==2.8.2
tenacity==8.2.3
markdownify==0.11.6
tiktoken==0.5.1
